+++ geotrek/maintenance/templates/maintenance/sql/post_90_defaults.sql
@@ -1,7 +1,7 @@
 -- Intervention
 ---------------
 -- name
-ALTER TABLE maintenance_intervention ALTER COLUMN "date" SET DEFAULT now();
+ALTER TABLE maintenance_intervention ALTER COLUMN "begin_date" SET DEFAULT now();
 ALTER TABLE maintenance_intervention ALTER COLUMN subcontracting SET DEFAULT FALSE;
 ALTER TABLE maintenance_intervention ALTER COLUMN width SET DEFAULT 0.0;
 ALTER TABLE maintenance_intervention ALTER COLUMN height SET DEFAULT 0.0;