+++ geotrek/signage/filters.py
@@ -46,7 +46,7 @@ class SignageFilterSet(AltimetryPointFilterSet, ValidTopologyFilterSet, ZoningFi
 
     def filter_intervention_year(self, qs, name, value):
         signage_ct = ContentType.objects.get_for_model(Signage)
-        interventions = Intervention.objects.filter(target_type=signage_ct, date__year__in=value) \
+        interventions = Intervention.objects.filter(target_type=signage_ct, begin_date__year__in=value) \
             .values_list('target_id', flat=True)
         return qs.filter(id__in=interventions).distinct()